pyproject.toml
[project]
name = "my_project"
version = "0.1.0"
description = ""
authors = [
    {name = "JeonDoGyun",email = "dodogugu2632@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
